add confirm modal before action

+++ components/post/delete-post.tsx
@@ -1,9 +1,30 @@
 "use client";
 
-import { Delete, Trash, Trash2 } from "lucide-react";
+import { Delete, EllipsisVertical, Trash, Trash2 } from "lucide-react";
 import { usePathname, useRouter } from "next/navigation";
+import { toast } from "sonner";
 
 import { deletePost } from "@/lib/actions/post";
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogPortal,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import { Button } from "../ui/button";
 
 interface Props {
   postId: string;
@@ -26,17 +47,50 @@ export const DeletePost = ({
   if (currentUserId !== authorId) return null;
 
   return (
-    <div className="flex items-center gap-x-2 cursor-pointer">
-      <Trash2
-        className="object-contain h-5 w-5"
-        onClick={async () => {
-          await deletePost(JSON.parse(postId), pathname);
-          if (!parentId || !isComment) {
-            router.push("/");
-          }
-        }}
-      />
-      <p>Delete Post</p>
-    </div>
+    <AlertDialog>
+      <DropdownMenu>
+        <DropdownMenuTrigger asChild>
+          <Button variant="ghost" className="rounded-full" size="icon">
+            <EllipsisVertical className="h-5 w-5 cursor-pointer" />
+          </Button>
+        </DropdownMenuTrigger>
+
+        <DropdownMenuContent align="end">
+          <AlertDialogTrigger>
+            <DropdownMenuItem>
+              <div className="flex items-center gap-x-2 cursor-pointer">
+                <Trash2 className="object-contain h-5 w-5" />
+                <p>Delete Post</p>
+              </div>
+            </DropdownMenuItem>
+          </AlertDialogTrigger>
+        </DropdownMenuContent>
+      </DropdownMenu>
+
+      <AlertDialogPortal>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+            <AlertDialogDescription>
+              This action cannot be undone. This will permanently delete post.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={async () => {
+                await deletePost(JSON.parse(postId), pathname);
+                toast.success("Post Deleted!");
+                if (!parentId || !isComment) {
+                  router.push("/");
+                }
+              }}
+            >
+              Continue
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialogPortal>
+    </AlertDialog>
   );
 };
